pyproject.toml
[project]
name = "swarm-agent"
version = "0.1.0"
description = "Decentralized Swarm Intelligence Agent Matrix - a self-organizing multi-agent system based on blackboard architecture"
authors = [
    {name = "zhaoJinPeng", email = "774461674@qq.com"}
]
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
keywords = ["agent", "multi-agent", "swarm", "llm", "mcp", "blackboard"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "python-dotenv",
    "fastapi (>=0.115.0)",
    "uvicorn (>=0.34.0)",
    "openai (>=1.0.0)",
]

[project.optional-dependencies]
mcp = ["mcp (>=1.0.0)"]

[project.urls]
Homepage = "https://github.com/manman-zjp/swarm-Agent"
Repository = "https://github.com/manman-zjp/swarm-Agent"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
